requirements.txt
rioxarray
xarray
xarray-spatial
osmnx
scipy
geopandas
rasterio
numpy
matplotlib
datashader
streamlit
folium
streamlit-folium
Pillow
requests
networkx
gdacs-api
plotly
stable-baselines3
gymnasium